+++ src/pages/admin/index.tsx
@@ -1,10 +1,11 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import { Container } from '@mui/material';
+import { Container, Toolbar } from '@mui/material';
 
 import { useAppSelector } from '../../store/hooks';
 import { authSelector } from '../../store/helpers/auth';
 import MainAdmin from '../../component/admin';
+import MainHeader from '../../layout/header/MainHeader';
 
 const Admin = () => {
    const { fields } = useAppSelector((state) => authSelector(state));
@@ -23,6 +24,9 @@ const Admin = () => {
 
    return (
       <Container>
+         <MainHeader />
+         <Toolbar />
+
          <MainAdmin />
       </Container>
    );
